requirements.txt
# APScheduler==3.10.4
Django==5.1
django-apscheduler==0.6.2
django-import-export==4.1.1
djangorestframework==3.15.2
drf-spectacular==0.27.2